+++ scripts/dev/lint.sh
@@ -97,4 +97,3 @@ else
     echo ""
     exit 1
 fi
-